+++ web/js/ui/account_upgrade/AccountUpgradeBarView.tsx
@@ -4,12 +4,16 @@ import {Button} from "reactstrap";
 import {AccountUpgrades, AccountUsage} from "../../accounts/AccountUpgrades";
 import {AccountPlan} from "../../accounts/Account";
 import {RendererAnalytics} from "../../ga/RendererAnalytics";
+import {Arrays} from "../../util/Arrays";
 
 const log = Logger.create();
+
+const MESSAGE = createRandomizedUpgradeMessage();
 
 interface UpgradeRequiredProps {
     readonly planRequired?: AccountPlan;
 }
+
 const UpgradeRequired = (props: UpgradeRequiredProps) => {
 
     RendererAnalytics.event({category: 'upgrade', action: 'triggered-upgrade-required'});
@@ -44,6 +48,42 @@ const UpgradeRequired = (props: UpgradeRequiredProps) => {
 
 };
 
+
+const GoPremium = (props: UpgradeRequiredProps) => {
+
+    RendererAnalytics.event({category: 'upgrade', action: 'triggered-upgrade-required'});
+
+    const onClick = () => {
+        RendererAnalytics.event({category: 'upgrade', action: 'clicked-button-to-plans'});
+        document.location.hash = 'plans';
+    };
+
+    return <div className="mt-1 mb-1 p-1 rounded"
+                style={{
+                    backgroundColor: '#ffcccc',
+                    fontWeight: 'bold'
+                }}>
+
+        <Button color="primary"
+                size="sm"
+                style={{fontWeight: 'bold'}}
+                onClick={() => onClick()}>
+
+            <i className="fas fa-certificate"/>
+            &nbsp;
+            Go Premium!
+
+        </Button>
+
+        <span className="ml-1">
+            {MESSAGE}
+        </span>
+
+    </div>;
+
+};
+
+
 const NullComponent = () => {
     return <div/>;
 };
@@ -62,17 +102,21 @@ export class AccountUpgradeBarView extends React.Component<IProps, IState> {
 
         const {plan, accountUsage} = this.props;
 
-        if (! plan || ! accountUsage) {
-            return <NullComponent/>;
+        if (plan && accountUsage) {
+
+            const planRequiredForUpgrade = AccountUpgrades.upgradeRequired(plan, accountUsage);
+
+            if (planRequiredForUpgrade) {
+                return <UpgradeRequired planRequired={planRequiredForUpgrade}/>;
+            }
+
         }
 
-        const planRequired = AccountUpgrades.upgradeRequired(plan, accountUsage);
-
-        if (! planRequired) {
+        if (! plan || plan === 'free') {
+            return <GoPremium/>;
+        } else {
             return <NullComponent/>;
         }
-
-        return <UpgradeRequired planRequired={planRequired}/>;
 
     }
 
@@ -84,4 +128,19 @@ interface IProps {
 }
 
 interface IState {
+}
+
+function  createRandomizedUpgradeMessage() {
+
+    const messages = [
+        "Want a dark mode? How about ePub support? Go premium and support Polar development!",
+        "Premium users help support future Polar development and are generally pretty awesome.",
+        "Guess who else used Polar Premium? Einstein! You want to be like Einstein don't you?",
+        "It's scientifically proven that Polar premium adds 100 years to your life!",
+    ];
+
+    const randomized = Arrays.shuffle(...messages);
+
+    return Arrays.first(randomized);
+
 }